% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_element}
\alias{get_element}
\title{Get element from object}
\usage{
get_element(object, name)
}
\arguments{
\item{object}{object from which to extract element(s) or in which to replace element(s)}

\item{name}{A literal character string or a name (possibly backtick quoted) or an index}
}
\description{
Get element from vectors, matrices, arrays, lists, environments, reference class objects, etc.
}

