% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_missing_translations}
\alias{get_missing_translations}
\title{Get missing translations as a table}
\usage{
get_missing_translations()
}
\description{
Get missing translations as a table
}
\seealso{
Other translation.table: \code{\link{get_translation_table}},
  \code{\link{set_translation_table}}
}

