\name{getGoogleLanguages}
\alias{getGoogleLanguages}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Print Google Language Codes}
\description{This function prints the valid language Google language codes for use with the translate() function.}
\usage{
getGoogleLanguages()
}
\examples{
# print valid language codes
getGoogleLanguages()
}
