\name{getMicrosoftLanguages}
\alias{getMicrosoftLanguages}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Print Microsoft Language Codes}
\description{This function prints the valid language Microsoft language codes for use with the translate() function.}
\usage{
getMicrosoftLanguages()
}
\examples{
# print valid language codes
getMicrosoftLanguages()
}
