% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_in_missing_translations}
\alias{is_in_missing_translations}
\title{Is the string already in the table for missing translations?}
\usage{
is_in_missing_translations(string, source.language, object.name)
}
\arguments{
\item{string}{string in the source language}

\item{source.language}{source language (abbreviation) of the string}

\item{object.name}{name of the object to be translated, to identify the object in the translation table, makes each entry of the translation table unique}
}
\description{
Is the string already in the table for missing translations?
}

