% Generated by roxygen2 (4.0.1): do not edit by hand
\name{record_missing_translations}
\alias{record_missing_translations}
\title{Add missing translation string to a table of missing translations}
\usage{
record_missing_translations(string, source.language, object.name)
}
\arguments{
\item{string}{string in the source language}

\item{source.language}{source language (abbreviation) of the string}

\item{object.name}{name of the object to be translated, to identify the object in the translation table, makes each entry of the translation table unique}
}
\description{
Internal function to keep the strings that have no translation.
}

