% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranslateR.R
\name{translate}
\alias{translate}
\title{Translate DNA or RNA sequence to amino acid}
\usage{
translate(sequence)
}
\arguments{
\item{sequence}{a text string containg a DNA or RNA sequence.}
}
\value{
A text string that represents the translated amino acid sequence.
}
\description{
Translate DNA or RNA sequence to amino acid
}
\examples{
sequence <- "ATGCTGTAG"
translate(sequence)
}
