% Generated by roxygen2 (4.0.1): do not edit by hand
\name{translate_object}
\alias{translate_object}
\title{Translate any string in an object}
\usage{
translate_object(object, source.language = "sv", target.language = "en",
  object.name = NULL, all.names = FALSE, skip = c("required_aes",
  "objname", ".self", ".refClassDef"), verbose = FALSE, level = NULL)
}
\arguments{
\item{object}{object that contains strings to translate}

\item{source.language}{the language of the given string, column name in the translation table}

\item{target.language}{target language, column name in the translation table}

\item{object.name}{name of the object to be translated, to identify the object in the translation table, makes each entry of the translation table unique}

\item{all.names}{logical, should all elements be translated, hidden (names begining with '.') elements included? Passed to \code{\link{ls}}}

\item{skip}{character vector, containing names of object that should not be translated}

\item{verbose}{a logical. If TRUE, additional diagnostics are printed}

\item{level}{used for verbose purpose to keep track of the nesting in the object}
}
\description{
Translate any string in an object
}
\details{
If you would like to translate the strings in an object containing references to environments, e.g. a ggplot object or a reference class object, you should use \code{\link{copy_object}} to copy the original object. Otherwise the original object may not behave as you would expect.
}

