% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_translation_table}
\alias{get_translation_table}
\alias{set_translation_table}
\title{Set/get translation table}
\usage{
set_translation_table(translation.table)

get_translation_table()
}
\arguments{
\item{translation.table}{data frame with at least three columns: a source language, a target language, and object.name.}
}
\description{
Set/get translation table to use with \code{\link{get_translation}}
}
\details{
The three columns of the translation table must have the names of the source language, e.g. \code{sv}, the target language, e.g. \code{en}, and \code{object.name}. The \code{object.name} makes each row unique together with one of the language columns. You can provide as many language columns as you need.
}
\seealso{
Other translation.table: \code{\link{get_missing_translations}}

Other translation.table: \code{\link{get_missing_translations}}
}

